# swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "Ecommerce Prisma Node Typescript API",
    "description": "Documentation automatically generated by the <b>swagger-autogen</b> module."
  },
  "host": "localhost:3000",
  "basePath": "/",
  "tags": [
    {
      "name": "User",
      "description": "Endpoints"
    }
  ],
  "schemes": [
    "http",
    "https"
  ],
  "securityDefinitions": {
    "apiKeyAuth": {
      "type": "authorization",
      "in": "header",
      "name": "authorization",
      "description": "Bearer Token"
    }
  },
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/auth/login": {
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "email": {
                  "example": "any"
                },
                "password": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {}
      }
    },
    "/auth/signup": {
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "email": {
                  "example": "any"
                },
                "name": {
                  "example": "any"
                },
                "dob": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {}
      }
    },
    "/auth/account-activation/{token}": {
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "token",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "password": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {}
      }
    },
    "/auth/reset-password": {
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "email": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {}
      }
    },
    "/auth/update-password/{token}": {
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "token",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "password": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {}
      }
    },
    "/user/profile": {
      "get": {
        "description": "",
        "parameters": [],
        "responses": {}
      }
    },
    "/user/preference": {
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "preferences": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {}
      }
    },
    "/preference/": {
      "get": {
        "description": "",
        "parameters": [],
        "responses": {}
      }
    }
  },
  "definitions": {
    "User": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string",
          "example": "Jhon Doe"
        },
        "email": {
          "type": "string",
          "example": "jhondoe@gmail.com"
        },
        "password": {
          "type": "string",
          "example": "#@!!!"
        },
        "dob": {
          "type": "string",
          "example": "2000-05-20"
        }
      },
      "required": [
        "name",
        "email"
      ]
    },
    "Preferences": {
      "type": "object",
      "properties": {
        "preference": {
          "type": "string",
          "example": "Furniture"
        }
      },
      "required": [
        "preference"
      ]
    },
    "UserPreferences": {
      "type": "object",
      "properties": {
        "userId": {
          "$ref": "#/definitions/User"
        },
        "preferenceId": {
          "$ref": "#/definitions/Preferences"
        }
      },
      "required": [
        "userId",
        "preferenceId"
      ]
    }
  }
}